+++ Packages/com.mimylab.fukuroudon/Runtime/ComponentIconPath.cs
@@ -1,0 +1,9 @@
+//using UnityEngine;
+
+namespace MimyLab.FukuroUdon
+{
+    static class ComponentIconPath
+    {
+        public const string FukuroUdon = "Packages/com.mimylab.fukuroudon/Icons/1f989.png";
+    }
+}