+++ Packages/com.mimylab.fukuroudon/Runtime/ManualObjectSync/Scripts/MOSUpdateManager.cs
@@ -4,7 +4,7 @@ Released under the MIT license
 https://opensource.org/licenses/mit-license.php
 */
 
-namespace MimyLab
+namespace MimyLab.FukuroUdon
 {
     using UdonSharp;
     using UnityEngine;
@@ -21,6 +21,7 @@ namespace MimyLab
     using UdonSharpEditor;
 #endif
 
+    [Icon(ComponentIconPath.FukuroUdon)]
     [UdonBehaviourSyncMode(BehaviourSyncMode.None)]
     public class MOSUpdateManager : UdonSharpBehaviour
     {