+++ Packages/com.mimylab.fukuroudon/Runtime/PlayerAudioMaster/Scripts/PARListController.cs
@@ -4,7 +4,7 @@ Released under the MIT license
 https://opensource.org/licenses/mit-license.php
 */
 
-namespace MimyLab
+namespace MimyLab.FukuroUdon
 {
     using UdonSharp;
     using UnityEngine;
@@ -19,6 +19,7 @@ namespace MimyLab
         ReleaseAll
     }
 
+    [Icon(ComponentIconPath.FukuroUdon)]
     [AddComponentMenu("Fukuro Udon/PlayerAudio Master/PAR List Controller")]
     [UdonBehaviourSyncMode(BehaviourSyncMode.None)]
     public class PARListController : UdonSharpBehaviour