+++ Packages/com.mimylab.fukuroudon/Runtime/PlayerAudioMaster/Scripts/PlayerAudioRegulatorList.cs
@@ -4,7 +4,7 @@ Released under the MIT license
 https://opensource.org/licenses/mit-license.php
 */
 
-namespace MimyLab
+namespace MimyLab.FukuroUdon
 {
     using UdonSharp;
     using UnityEngine;