+++ Packages/com.mimylab.fukuroudon/Runtime/SwivelChair/Scripts/SCKeyInputManager.cs
@@ -4,13 +4,14 @@ Released under the MIT license
 https://opensource.org/licenses/mit-license.php
 */
 
-namespace MimyLab
+namespace MimyLab.FukuroUdon
 {
     using UdonSharp;
     using UnityEngine;
     //using VRC.SDKBase;
     //using VRC.Udon;
 
+    [Icon(ComponentIconPath.FukuroUdon)]
     [DefaultExecutionOrder(10)]
     [UdonBehaviourSyncMode(BehaviourSyncMode.Manual)]
     public class SCKeyInputManager : UdonSharpBehaviour