+++ Packages/com.mimylab.fukuroudon/CHANGELOG.md
@@ -6,6 +6,13 @@
 
 利用手順は[こちら](https://github.com/mimyquality/FukuroUdon/wiki)からご確認ください。
 
+## [2.0.0-beta] - 2024/8/30
+
+- **Changed**
+  - サポートバージョンをUnity2022.3、VRCSDK 3.5.0以上に引き上げました。
+  - 全てのU#スクリプトの名前空間を `MimyLab` から `MimyLab.FukuroUdon` に変更しました。
+  - 全てのコンポーネントにアイコンを実装しました。
+
 ## [1.24.0] - 2024/8/3
 
 - **Added**
@@ -509,6 +516,7 @@
     - 0.2.7バージョンで、later-joiner視点でjoin直後が非アクティブだと同期しなかったのを修正しました。
     - 他人がピックアップした際にオブジェクトが一瞬跳ねる現象を低減しました。
 
+[2.0.0-beta]: https://github.com/mimyquality/FukuroUdon/releases/tag/2.0.0-beta
 [1.24.0]: https://github.com/mimyquality/FukuroUdon/releases/tag/1.24.0
 [1.23.0]: https://github.com/mimyquality/FukuroUdon/releases/tag/1.23.0
 [1.22.1]: https://github.com/mimyquality/FukuroUdon/releases/tag/1.22.1